+++ NEWS.md
@@ -9,4 +9,10 @@
 2020-12-19
 
 * Added a `NEWS.md` file to track changes to the package.
-* Updated the function `check()`. Now it detach packages which are attached by an user before check a solution and attach it again after check.
+* Updated the function `check_solution()`. Now it detaches packages which are attached by an user before check a solution and attach it again after check.
+
+# 0.3.0
+
+2021-05-27
+
+* Updated the function `check_solution()`. Now it checks for restricted functions. Also the new feature test mode is added.